settings_account: Allow passing custom profile fields as argument.
`initialize_custom_user_type_fields` searched for user type fields using the
`page_params.custom_profile_fields` list. To be able to re-use this function for
displaying user type fields in the small profile, this function has been
refactored to accept a list of profile fields to look in, for user type fields.